Move out colormap into lib
